Cover leading newline export text
